[WIP] sdig: Add DNS over TLS support
