+++ pdns/dnsdistdist/configure.ac
@@ -66,15 +66,15 @@ AM_CONDITIONAL([HAVE_LIBSSL], [false])
 
 PDNS_CHECK_LIBCRYPTO
 
-DNSDIST_ENABLE_DNS_OVER_TLS
+PDNS_ENABLE_DNS_OVER_TLS
 DNSDIST_ENABLE_DNS_OVER_HTTPS
 
 AS_IF([test "x$enable_dns_over_tls" != "xno" -o "x$enable_dns_over_https" != "xno"], [
-  DNSDIST_WITH_LIBSSL
+  PDNS_WITH_LIBSSL
 ])
 
 AS_IF([test "x$enable_dns_over_tls" != "xno"], [
-  DNSDIST_WITH_GNUTLS
+  PDNS_WITH_GNUTLS
 
   AS_IF([test "x$HAVE_GNUTLS" != "x1" -a "x$HAVE_LIBSSL" != "x1"], [
     AC_MSG_ERROR([DNS over TLS support requested but neither GnuTLS nor OpenSSL are available])
