+++ pdns/dnsdistdist/m4/pdns_with_libssl.m4
@@ -1,0 +1,1 @@
+../../../m4/pdns_with_libssl.m4
